Replace the Signer with access to signing via new traits SignableZone and SignableZoneInPlace, removing the need for state to be held, especially as some of that state was better associated with the zone being signed than with the signer.
Introduce a new structure under sign/ to hold types extracted by refactoring the long records.rs into separate modules.

Add generalized RecordSet::update_data() instead of the very specific replace_soa() and replace_rrsign_for_apex_zonemd() functions.
