+++ src/sign/records.rs
@@ -1,40 +1,19 @@
 //! Actual signing.
 use core::cmp::Ordering;
 use core::convert::From;
-use core::fmt::Display;
-use core::marker::PhantomData;
-use core::ops::Deref;
+use core::iter::Extend;
+use core::marker::{PhantomData, Send};
 use core::slice::Iter;
 
-use std::boxed::Box;
-use std::collections::HashSet;
-use std::fmt::Debug;
-use std::hash::Hash;
-use std::string::{String, ToString};
 use std::vec::Vec;
 use std::{fmt, slice};
 
-use bytes::Bytes;
-use octseq::builder::{EmptyBuilder, FromBuilder, OctetsBuilder, Truncate};
-use octseq::{FreezeBuilder, OctetsFrom, OctetsInto};
-use tracing::{debug, enabled, trace, Level};
-
 use crate::base::cmp::CanonicalOrd;
-use crate::base::iana::{Class, Nsec3HashAlg, Rtype};
-use crate::base::name::{ToLabelIter, ToName};
-use crate::base::rdata::{ComposeRecordData, RecordData};
+use crate::base::iana::{Class, Rtype};
+use crate::base::name::ToName;
+use crate::base::rdata::RecordData;
 use crate::base::record::Record;
-use crate::base::{Name, NameBuilder, Ttl};
-use crate::rdata::dnssec::{ProtoRrsig, RtypeBitmap, RtypeBitmapBuilder};
-use crate::rdata::nsec3::{Nsec3Salt, OwnerHash};
-use crate::rdata::{
-    Dnskey, Nsec, Nsec3, Nsec3param, Rrsig, Soa, ZoneRecordData,
-};
-use crate::sign::{SignRaw, SigningKey};
-use crate::utils::base32;
-use crate::validate::{nsec3_hash, Nsec3HashError};
-use crate::zonetree::types::StoredRecordData;
-use crate::zonetree::StoredName;
+use crate::base::Ttl;
 
 //------------ Sorter --------------------------------------------------------
 
@@ -222,6 +201,33 @@ where
         self.rrsets().find(|rrset| rrset.rtype() == Rtype::SOA)
     }
 
+    /// Update the data of an existing record.
+    ///
+    /// Allowing records to be mutated in-place would not be safe because it
+    /// could invalidate the sort order so no general method to mutate the
+    /// records is provided.
+    ///
+    /// This method offers a limited ability to mutate records in-place
+    /// however because it only permits mutating of the resource record data
+    /// of an existing record which doesn't impact the sort order because the
+    /// data is not part of the sort key.
+    pub fn update_data<F>(&mut self, matcher: F, new_data: D)
+    where
+        F: Fn(&Record<N, D>) -> bool,
+    {
+        if let Some(rr) = self.records.iter_mut().find(|rr| matcher(rr)) {
+            *rr.data_mut() = new_data;
+        }
+    }
+
+    pub fn len(&self) -> usize {
+        self.records.len()
+    }
+
+    pub fn is_empty(&self) -> bool {
+        self.records.is_empty()
+    }
+
     pub fn iter(&self) -> Iter<'_, Record<N, D>> {
         self.records.iter()
     }
@@ -230,48 +236,21 @@ where
         self.records.as_slice()
     }
 
+    pub(super) fn as_mut_slice(&mut self) -> &mut [Record<N, D>] {
+        self.records.as_mut_slice()
+    }
+
     pub fn into_inner(self) -> Vec<Record<N, D>> {
         self.records
     }
 }
 
-impl<N: Send + ToName, S: Sorter> SortedRecords<N, StoredRecordData, S> {
-    pub fn replace_soa(&mut self, new_soa: Soa<StoredName>) {
-        if let Some(soa_rrset) = self
-            .records
-            .iter_mut()
-            .find(|rrset| rrset.rtype() == Rtype::SOA)
-        {
-            if let ZoneRecordData::Soa(current_soa) = soa_rrset.data_mut() {
-                *current_soa = new_soa;
-            }
-        }
-    }
-
-    pub fn replace_rrsig_for_apex_zonemd(
-        &mut self,
-        new_rrsig: Rrsig<Bytes, StoredName>,
-        apex: &FamilyName<StoredName>,
-    ) {
-        if let Some(zonemd_rrsig) = self.records.iter_mut().find(|record| {
-            if record.rtype() == Rtype::RRSIG
-                && record.owner().name_cmp(&apex.owner()) == Ordering::Equal
-            {
-                if let ZoneRecordData::Rrsig(rrsig) = record.data() {
-                    if rrsig.type_covered() == Rtype::ZONEMD {
-                        return true;
-                    }
-                }
-            }
-            false
-        }) {
-            if let ZoneRecordData::Rrsig(current_rrsig) =
-                zonemd_rrsig.data_mut()
-            {
-                *current_rrsig = new_rrsig;
-            }
-        }
-    }
+impl<N, D, S> SortedRecords<N, D, S>
+where
+    N: Send,
+    D: Send,
+    S: Sorter,
+{
 }
 
 impl<N, D, S> SortedRecords<N, D, S>
@@ -281,528 +260,6 @@ where
     S: Sorter,
     SortedRecords<N, D>: From<Vec<Record<N, D>>>,
 {
-    pub fn nsecs<Octets>(
-        &self,
-        apex: &FamilyName<N>,
-        ttl: Ttl,
-        assume_dnskeys_will_be_added: bool,
-    ) -> Vec<Record<N, Nsec<Octets, N>>>
-    where
-        N: ToName + Clone + PartialEq,
-        D: RecordData,
-        Octets: FromBuilder,
-        Octets::Builder: EmptyBuilder + Truncate + AsRef<[u8]> + AsMut<[u8]>,
-        <Octets::Builder as OctetsBuilder>::AppendError: Debug,
-    {
-        let mut res = Vec::new();
-
-        // The owner name of a zone cut if we currently are at or below one.
-        let mut cut: Option<FamilyName<N>> = None;
-
-        let mut families = self.families();
-
-        // Since the records are ordered, the first family is the apex --
-        // we can skip everything before that.
-        families.skip_before(apex);
-
-        // Because of the next name thing, we need to keep the last NSEC
-        // around.
-        let mut prev: Option<(FamilyName<N>, RtypeBitmap<Octets>)> = None;
-
-        // We also need the apex for the last NSEC.
-        let apex_owner = families.first_owner().clone();
-
-        for family in families {
-            // If the owner is out of zone, we have moved out of our zone and
-            // are done.
-            if !family.is_in_zone(apex) {
-                break;
-            }
-
-            // If the family is below a zone cut, we must ignore it.
-            if let Some(ref cut) = cut {
-                if family.owner().ends_with(cut.owner()) {
-                    continue;
-                }
-            }
-
-            // A copy of the family name. We’ll need it later.
-            let name = family.family_name().cloned();
-
-            // If this family is the parent side of a zone cut, we keep the
-            // family name for later. This also means below that if
-            // `cut.is_some()` we are at the parent side of a zone.
-            cut = if family.is_zone_cut(apex) {
-                Some(name.clone())
-            } else {
-                None
-            };
-
-            if let Some((prev_name, bitmap)) = prev.take() {
-                res.push(prev_name.into_record(
-                    ttl,
-                    Nsec::new(name.owner().clone(), bitmap),
-                ));
-            }
-
-            let mut bitmap = RtypeBitmap::<Octets>::builder();
-            // RFC 4035 section 2.3:
-            //   "The type bitmap of every NSEC resource record in a signed
-            //    zone MUST indicate the presence of both the NSEC record
-            //    itself and its corresponding RRSIG record."
-            bitmap.add(Rtype::RRSIG).unwrap();
-            if assume_dnskeys_will_be_added && family.owner() == &apex_owner {
-                // Assume there's gonna be a DNSKEY.
-                bitmap.add(Rtype::DNSKEY).unwrap();
-            }
-            bitmap.add(Rtype::NSEC).unwrap();
-            for rrset in family.rrsets() {
-                // RFC 4034 section 4.1.2: (and also RFC 4035 section 2.3)
-                //   "The bitmap for the NSEC RR at a delegation point
-                //    requires special attention.  Bits corresponding to the
-                //    delegation NS RRset and any RRsets for which the parent
-                //    zone has authoritative data MUST be set; bits
-                //    corresponding to any non-NS RRset for which the parent
-                //    is not authoritative MUST be clear."
-                if cut.is_none()
-                    || matches!(rrset.rtype(), Rtype::NS | Rtype::DS)
-                {
-                    // RFC 4034 section 4.1.2:
-                    //   "Bits representing pseudo-types MUST be clear, as
-                    //    they do not appear in zone data."
-                    //
-                    // TODO: Should this check be moved into
-                    // RtypeBitmapBuilder itself?
-                    if !rrset.rtype().is_pseudo() {
-                        bitmap.add(rrset.rtype()).unwrap()
-                    }
-                }
-            }
-
-            prev = Some((name, bitmap.finalize()));
-        }
-        if let Some((prev_name, bitmap)) = prev {
-            res.push(
-                prev_name.into_record(ttl, Nsec::new(apex_owner, bitmap)),
-            );
-        }
-        res
-    }
-
-    /// Generate [RFC5155] NSEC3 and NSEC3PARAM records for this record set.
-    ///
-    /// This function does NOT enforce use of current best practice settings,
-    /// as defined by [RFC 5155], [RFC 9077] and [RFC 9276] which state that:
-    ///
-    /// - The `ttl` should be the _"lesser of the MINIMUM field of the zone
-    ///   SOA RR and the TTL of the zone SOA RR itself"_.
-    ///
-    /// - The `params` should be set to _"SHA-1, no extra iterations, empty
-    ///   salt"_ and zero flags. See [`Nsec3param::default()`].
-    ///
-    /// [RFC 5155]: https://www.rfc-editor.org/rfc/rfc5155.html
-    /// [RFC 9077]: https://www.rfc-editor.org/rfc/rfc9077.html
-    /// [RFC 9276]: https://www.rfc-editor.org/rfc/rfc9276.html
-    // TODO: Move to Signer and do HashProvider = OnDemandNsec3HashProvider
-    // TODO: Does it make sense to take both Nsec3param AND HashProvider as input?
-    pub fn nsec3s<Octets, OctetsMut, HashProvider>(
-        &self,
-        apex: &FamilyName<N>,
-        ttl: Ttl,
-        params: Nsec3param<Octets>,
-        opt_out: Nsec3OptOut,
-        assume_dnskeys_will_be_added: bool,
-        hash_provider: &mut HashProvider,
-    ) -> Result<Nsec3Records<N, Octets>, Nsec3HashError>
-    where
-        N: ToName + Clone + From<Name<Octets>> + Display + Ord + Hash,
-        N: From<Name<<OctetsMut as FreezeBuilder>::Octets>>,
-        D: RecordData + From<Nsec3<Octets>>,
-        Octets: Send + FromBuilder + OctetsFrom<Vec<u8>> + Clone + Default,
-        Octets::Builder: EmptyBuilder + Truncate + AsRef<[u8]> + AsMut<[u8]>,
-        <Octets::Builder as OctetsBuilder>::AppendError: Debug,
-        OctetsMut: OctetsBuilder
-            + AsRef<[u8]>
-            + AsMut<[u8]>
-            + EmptyBuilder
-            + FreezeBuilder,
-        <OctetsMut as FreezeBuilder>::Octets: AsRef<[u8]>,
-        HashProvider: Nsec3HashProvider<N, Octets>,
-        Nsec3<Octets>: Into<D>,
-    {
-        // TODO:
-        //   - Handle name collisions? (see RFC 5155 7.1 Zone Signing)
-        //   - RFC 5155 section 2 Backwards compatibility:
-        //     Reject old algorithms? if not, map 3 to 6 and 5 to 7, or reject
-        //     use of 3 and 5?
-
-        // RFC 5155 7.1 step 2:
-        //   "If Opt-Out is being used, set the Opt-Out bit to one."
-        let mut nsec3_flags = params.flags();
-        if matches!(
-            opt_out,
-            Nsec3OptOut::OptOut | Nsec3OptOut::OptOutFlagsOnly
-        ) {
-            // Set the Opt-Out flag.
-            nsec3_flags |= 0b0000_0001;
-        }
-
-        // RFC 5155 7.1 step 5: "Sort the set of NSEC3 RRs into hash order."
-        // We store the NSEC3s as we create them and sort them afterwards.
-        let mut nsec3s = Vec::<Record<N, Nsec3<Octets>>>::new();
-
-        let mut ents = Vec::<N>::new();
-
-        // The owner name of a zone cut if we currently are at or below one.
-        let mut cut: Option<FamilyName<N>> = None;
-
-        let mut families = self.families();
-
-        // Since the records are ordered, the first family is the apex --
-        // we can skip everything before that.
-        families.skip_before(apex);
-
-        // We also need the apex for the last NSEC.
-        let apex_owner = families.first_owner().clone();
-        let apex_label_count = apex_owner.iter_labels().count();
-
-        let mut last_nent_stack: Vec<N> = vec![];
-
-        for family in families {
-            trace!("Family: {}", family.family_name().owner());
-
-            // If the owner is out of zone, we have moved out of our zone and
-            // are done.
-            if !family.is_in_zone(apex) {
-                debug!(
-                    "Stopping NSEC3 generation at out-of-zone family {}",
-                    family.family_name().owner()
-                );
-                break;
-            }
-
-            // If the family is below a zone cut, we must ignore it. As the
-            // RRs are required to be sorted all RRs below a zone cut should
-            // be encountered after the cut itself.
-            if let Some(ref cut) = cut {
-                if family.owner().ends_with(cut.owner()) {
-                    debug!(
-                        "Excluding family {} as it is below a zone cut",
-                        family.family_name().owner()
-                    );
-                    continue;
-                }
-            }
-
-            // A copy of the family name. We’ll need it later.
-            let name = family.family_name().cloned();
-
-            // If this family is the parent side of a zone cut, we keep the
-            // family name for later. This also means below that if
-            // `cut.is_some()` we are at the parent side of a zone.
-            cut = if family.is_zone_cut(apex) {
-                trace!(
-                    "Zone cut detected at family {}",
-                    family.family_name().owner()
-                );
-                Some(name.clone())
-            } else {
-                None
-            };
-
-            // RFC 5155 7.1 step 2:
-            //   "If Opt-Out is being used, owner names of unsigned
-            //    delegations MAY be excluded."
-            // Note that:
-            //   - A "delegation inherently happens at a zone cut" (RFC 9499).
-            //   - An "unsigned delegation" aka an "insecure delegation" is a
-            //     "signed name containing a delegation (NS RRset), but
-            //     lacking a DS RRset, signifying a delegation to an unsigned
-            //     subzone" (RFC 9499).
-            // So we need to check for whether Opt-Out is being used at a zone
-            // cut that lacks a DS RR. We determine whether or not a DS RR is
-            // present even when Opt-Out is not being used because we also
-            // need to know there at a later step.
-            let has_ds = family.records().any(|rec| rec.rtype() == Rtype::DS);
-            if opt_out == Nsec3OptOut::OptOut && cut.is_some() && !has_ds {
-                debug!("Excluding family {} as it is an insecure delegation (lacks a DS RR) and opt-out is enabled",family.family_name().owner());
-                continue;
-            }
-
-            // RFC 5155 7.1 step 4:
-            //   "If the difference in number of labels between the apex and
-            //    the original owner name is greater than 1, additional NSEC3
-            //    RRs need to be added for every empty non-terminal between
-            //    the apex and the original owner name."
-            let mut last_nent_distance_to_apex = 0;
-            let mut last_nent = None;
-            while let Some(this_last_nent) = last_nent_stack.pop() {
-                if name.owner().ends_with(&this_last_nent) {
-                    last_nent_distance_to_apex =
-                        this_last_nent.iter_labels().count()
-                            - apex_label_count;
-                    last_nent = Some(this_last_nent);
-                    break;
-                }
-            }
-            let distance_to_root = name.owner().iter_labels().count();
-            let distance_to_apex = distance_to_root - apex_label_count;
-            if distance_to_apex > last_nent_distance_to_apex {
-                trace!(
-                    "Possible ENT detected at family {}",
-                    family.family_name().owner()
-                );
-
-                // Are there any empty nodes between this node and the apex?
-                // The zone file records are already sorted so if all of the
-                // parent labels had records at them, i.e. they were non-empty
-                // then non_empty_label_count would be equal to label_distance.
-                // If it is less that means there are ENTs between us and the
-                // last non-empty label in our ancestor path to the apex.
-
-                // Walk from the owner name down the tree of labels from the
-                // last known non-empty non-terminal label, extending the name
-                // each time by one label until we get to the current name.
-
-                // Given a.b.c.mail.example.com where:
-                //   - example.com is the apex owner
-                //   - mail.example.com was the last non-empty non-terminal
-                // This loop will construct the names:
-                //   - c.mail.example.com
-                //   - b.c.mail.example.com
-                // It will NOT construct the last name as that will be dealt
-                // with in the next outer loop iteration.
-                //   - a.b.c.mail.example.com
-                let distance = distance_to_apex - last_nent_distance_to_apex;
-                for n in (1..=distance - 1).rev() {
-                    let rev_label_it = name.owner().iter_labels().skip(n);
-
-                    // Create next longest ENT name.
-                    let mut builder = NameBuilder::<OctetsMut>::new();
-                    for label in rev_label_it.take(distance_to_apex - n) {
-                        builder.append_label(label.as_slice()).unwrap();
-                    }
-                    let name =
-                        builder.append_origin(&apex_owner).unwrap().into();
-
-                    if let Err(pos) = ents.binary_search(&name) {
-                        debug!("Found ENT at {name}");
-                        ents.insert(pos, name);
-                    }
-                }
-            }
-
-            // Create the type bitmap.
-            let mut bitmap = RtypeBitmap::<Octets>::builder();
-
-            // Authoritative RRsets will be signed by `sign()` so add the
-            // expected future RRSIG type now to the NSEC3 Type Bitmap we are
-            // constructing.
-            //
-            // RFC 4033 section 2:
-            // 2.  Definitions of Important DNSSEC Terms
-            //    Authoritative RRset: Within the context of a particular
-            //       zone, an RRset is "authoritative" if and only if the
-            //       owner name of the RRset lies within the subset of the
-            //       name space that is at or below the zone apex and at or
-            //       above the cuts that separate the zone from its children,
-            //       if any.  All RRsets at the zone apex are authoritative,
-            //       except for certain RRsets at this domain name that, if
-            //       present, belong to this zone's parent.  These RRset could
-            //       include a DS RRset, the NSEC RRset referencing this DS
-            //       RRset (the "parental NSEC"), and RRSIG RRs associated
-            //       with these RRsets, all of which are authoritative in the
-            //       parent zone.  Similarly, if this zone contains any
-            //       delegation points, only the parental NSEC RRset, DS
-            //       RRsets, and any RRSIG RRs associated with these RRsets
-            //       are authoritative for this zone.
-            if cut.is_none() || has_ds {
-                trace!("Adding RRSIG to the bitmap as the RRSET is authoritative (not at zone cut or has a DS RR)");
-                bitmap.add(Rtype::RRSIG).unwrap();
-            }
-
-            // RFC 5155 7.1 step 3:
-            //   "For each RRSet at the original owner name, set the
-            //    corresponding bit in the Type Bit Maps field."
-            //
-            // Note: When generating NSEC RRs (not NSEC3 RRs) RFC 4035 makes
-            // it clear that non-authoritative RRs should not be represented
-            // in the Type Bitmap but for NSEC3 generation that's less clear.
-            //
-            // RFC 4035 section 2.3:
-            // 2.3.  Including NSEC RRs in a Zone
-            //   ...
-            //   "bits corresponding to any non-NS RRset for which the parent
-            //   is not authoritative MUST be clear."
-            //
-            // RFC 5155 section 7.1:
-            // 7.1.  Zone Signing
-            //   ...
-            //   "o  The Type Bit Maps field of every NSEC3 RR in a signed
-            //       zone MUST indicate the presence of all types present at
-            //       the original owner name, except for the types solely
-            //       contributed by an NSEC3 RR itself.  Note that this means
-            //       that the NSEC3 type itself will never be present in the
-            //       Type Bit Maps."
-            //
-            // Thus the rules for the types to include in the Type Bitmap for
-            // NSEC RRs appear to be different for NSEC3 RRs. However, in
-            // practice common tooling implementations exclude types from the
-            // NSEC3 which are non-authoritative (e.g. glue and occluded
-            // records). One could argue that the following fragments of RFC
-            // 5155 support this:
-            //
-            // RFC 5155 section 7.1.
-            // 7.1.  Zone Signing
-            //   ...
-            //   "Other non-authoritative RRs are not represented by
-            //    NSEC3 RRs."
-            //   ...
-            //   "2.  For each unique original owner name in the zone add an
-            //   NSEC3 RR."
-            //
-            // (if one reads "in the zone" to exclude data occluded by a zone
-            // cut or glue records that are only authoritative in the child
-            // zone and not in the parent zone).
-            //
-            // RFC 4033 could also be interpreted as excluding
-            // non-authoritative data from DNSSEC and thus NSEC3:
-            //
-            // RFC 4033 section 9:
-            // 9.  Name Server Considerations
-            //   ...
-            //   "By itself, DNSSEC is not enough to protect the integrity of
-            //    an entire zone during zone transfer operations, as even a
-            //    signed zone contains some unsigned, nonauthoritative data if
-            //    the zone has any children."
-            //
-            // As such we exclude non-authoritative RRs from the NSEC3 Type
-            // Bitmap, with the EXCEPTION of the NS RR at a secure delegation
-            // as insecure delegations are explicitly included by RFC 5155:
-            //
-            // RFC 5155 section 7.1:
-            // 7.1.  Zone Signing
-            //   ...
-            //   "o  Each owner name within the zone that owns authoritative
-            //       RRSets MUST have a corresponding NSEC3 RR.  Owner names
-            //       that correspond to unsigned delegations MAY have a
-            //       corresponding NSEC3 RR."
-            for rrset in family.rrsets() {
-                if cut.is_none()
-                    || matches!(rrset.rtype(), Rtype::NS | Rtype::DS)
-                {
-                    // RFC 5155 section 3.2:
-                    //   "Bits representing Meta-TYPEs or QTYPEs as specified
-                    //    in Section 3.1 of [RFC2929] or within the range
-                    //    reserved for assignment only to QTYPEs and
-                    //    Meta-TYPEs MUST be set to 0, since they do not
-                    //    appear in zone data".
-                    //
-                    // TODO: Should this check be moved into
-                    // RtypeBitmapBuilder itself?
-                    if !rrset.rtype().is_pseudo() {
-                        trace!("Adding {} to the bitmap", rrset.rtype());
-                        bitmap.add(rrset.rtype()).unwrap();
-                    }
-                }
-            }
-
-            if distance_to_apex == 0 {
-                trace!("Adding NSEC3PARAM to the bitmap as we are at the apex and RRSIG RRs are expected to be added");
-                bitmap.add(Rtype::NSEC3PARAM).unwrap();
-                if assume_dnskeys_will_be_added {
-                    trace!("Adding DNSKEY to the bitmap as we are at the apex and DNSKEY RRs are expected to be added");
-                    bitmap.add(Rtype::DNSKEY).unwrap();
-                }
-            }
-
-            let rec: Record<N, Nsec3<Octets>> = Self::mk_nsec3(
-                name.owner(),
-                hash_provider,
-                params.hash_algorithm(),
-                nsec3_flags,
-                params.iterations(),
-                params.salt(),
-                &apex_owner,
-                bitmap,
-                ttl,
-            )?;
-
-            // Store the record by order of its owner name.
-            nsec3s.push(rec);
-
-            if let Some(last_nent) = last_nent {
-                last_nent_stack.push(last_nent);
-            }
-            last_nent_stack.push(name.owner().clone());
-        }
-
-        for name in ents {
-            // Create the type bitmap, empty for an ENT NSEC3.
-            let bitmap = RtypeBitmap::<Octets>::builder();
-
-            debug!("Generating NSEC3 RR for ENT at {name}");
-            let rec = Self::mk_nsec3(
-                &name,
-                hash_provider,
-                params.hash_algorithm(),
-                nsec3_flags,
-                params.iterations(),
-                params.salt(),
-                &apex_owner,
-                bitmap,
-                ttl,
-            )?;
-
-            // Store the record by order of its owner name.
-            nsec3s.push(rec);
-        }
-
-        // RFC 5155 7.1 step 7:
-        //   "In each NSEC3 RR, insert the next hashed owner name by using the
-        //    value of the next NSEC3 RR in hash order.  The next hashed owner
-        //    name of the last NSEC3 RR in the zone contains the value of the
-        //    hashed owner name of the first NSEC3 RR in the hash order."
-        trace!("Sorting NSEC3 RRs");
-        let mut nsec3s = SortedRecords::<N, Nsec3<Octets>, S>::from(nsec3s);
-        for i in 1..=nsec3s.records.len() {
-            // TODO: Detect duplicate hashes.
-            let next_i = if i == nsec3s.records.len() { 0 } else { i };
-            let cur_owner = nsec3s.records[next_i].owner();
-            let name: Name<Octets> = cur_owner.try_to_name().unwrap();
-            let label = name.iter_labels().next().unwrap();
-            let owner_hash = if let Ok(hash_octets) =
-                base32::decode_hex(&format!("{label}"))
-            {
-                OwnerHash::<Octets>::from_octets(hash_octets).unwrap()
-            } else {
-                OwnerHash::<Octets>::from_octets(name.as_octets().clone())
-                    .unwrap()
-            };
-            let last_rec = &mut nsec3s.records[i - 1];
-            let last_nsec3: &mut Nsec3<Octets> = last_rec.data_mut();
-            last_nsec3.set_next_owner(owner_hash.clone());
-        }
-
-        // RFC 5155 7.1 step 8:
-        //   "Finally, add an NSEC3PARAM RR with the same Hash Algorithm,
-        //    Iterations, and Salt fields to the zone apex."
-        let nsec3param = Record::new(
-            apex.owner().try_to_name::<Octets>().unwrap().into(),
-            Class::IN,
-            ttl,
-            params,
-        );
-
-        // RFC 5155 7.1 after step 8:
-        //   "If a hash collision is detected, then a new salt has to be
-        //    chosen, and the signing process restarted."
-        //
-        // Handled above.
-
-        Ok(Nsec3Records::new(nsec3s.records, nsec3param))
-    }
-
     pub fn write<W>(&self, target: &mut W) -> Result<(), fmt::Error>
     where
         N: fmt::Display,
@@ -848,58 +305,6 @@ where
         }
 
         Ok(())
-    }
-}
-
-/// Helper functions used to create NSEC3 records per RFC 5155.
-impl<N, D, S> SortedRecords<N, D, S>
-where
-    N: ToName + Send,
-    D: RecordData + CanonicalOrd + Send,
-    S: Sorter,
-{
-    #[allow(clippy::too_many_arguments)]
-    fn mk_nsec3<Octets, HashProvider>(
-        name: &N,
-        hash_provider: &mut HashProvider,
-        alg: Nsec3HashAlg,
-        flags: u8,
-        iterations: u16,
-        salt: &Nsec3Salt<Octets>,
-        _apex_owner: &N,
-        bitmap: RtypeBitmapBuilder<<Octets as FromBuilder>::Builder>,
-        ttl: Ttl,
-    ) -> Result<Record<N, Nsec3<Octets>>, Nsec3HashError>
-    where
-        N: ToName + From<Name<Octets>>,
-        Octets: FromBuilder + Clone + Default,
-        <Octets as FromBuilder>::Builder:
-            EmptyBuilder + AsRef<[u8]> + AsMut<[u8]> + Truncate,
-        Nsec3<Octets>: Into<D>,
-        HashProvider: Nsec3HashProvider<N, Octets>,
-    {
-        // let owner_name = mk_hashed_nsec3_owner_name(
-        //     name, alg, iterations, salt, apex_owner,
-        // )?;
-        let owner_name = hash_provider.get_or_create(name)?;
-
-        // RFC 5155 7.1. step 2:
-        //   "The Next Hashed Owner Name field is left blank for the moment."
-        // Create a placeholder next owner, we'll fix it later.
-        let placeholder_next_owner =
-            OwnerHash::<Octets>::from_octets(Octets::default()).unwrap();
-
-        // Create an NSEC3 record.
-        let nsec3 = Nsec3::new(
-            alg,
-            flags,
-            iterations,
-            salt.clone(),
-            placeholder_next_owner,
-            bitmap.finalize(),
-        );
-
-        Ok(Record::new(owner_name, Class::IN, ttl, nsec3))
     }
 }
 
@@ -954,25 +359,6 @@ where
         }
         S::sort_by(&mut self.records, CanonicalOrd::canonical_cmp);
         self.records.dedup();
-    }
-}
-
-//------------ Nsec3Records ---------------------------------------------------
-
-pub struct Nsec3Records<N, Octets> {
-    /// The NSEC3 records.
-    pub recs: Vec<Record<N, Nsec3<Octets>>>,
-
-    /// The NSEC3PARAM record.
-    pub param: Record<N, Nsec3param<Octets>>,
-}
-
-impl<N, Octets> Nsec3Records<N, Octets> {
-    pub fn new(
-        recs: Vec<Record<N, Nsec3<Octets>>>,
-        param: Record<N, Nsec3param<Octets>>,
-    ) -> Self {
-        Self { recs, param }
     }
 }
 
@@ -1086,7 +472,7 @@ pub struct Rrset<'a, N, D> {
 }
 
 impl<'a, N, D> Rrset<'a, N, D> {
-    fn new(slice: &'a [Record<N, D>]) -> Self {
+    pub fn new(slice: &'a [Record<N, D>]) -> Self {
         Rrset { slice }
     }
 
@@ -1134,7 +520,7 @@ pub struct RecordsIter<'a, N, D> {
 }
 
 impl<'a, N, D> RecordsIter<'a, N, D> {
-    fn new(slice: &'a [Record<N, D>]) -> Self {
+    pub fn new(slice: &'a [Record<N, D>]) -> Self {
         RecordsIter { slice }
     }
 
@@ -1261,831 +647,5 @@ where
         let (res, slice) = self.slice.split_at(end);
         self.slice = slice;
         Some(Rrset::new(res))
-    }
-}
-
-//------------ SigningError --------------------------------------------------
-
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-pub enum SigningError {
-    /// One or more keys does not have a signature validity period defined.
-    KeyLacksSignatureValidityPeriod,
-
-    /// TODO
-    OutOfMemory,
-
-    /// At least one key must be provided to sign with.
-    NoKeysProvided,
-
-    /// None of the provided keys were deemed suitable by the
-    /// [`SigningKeyUsageStrategy`] used.
-    NoSuitableKeysFound,
-}
-
-//------------ Nsec3OptOut ---------------------------------------------------
-
-/// The different types of NSEC3 opt-out.
-#[derive(Copy, Clone, Debug, Default, Eq, PartialEq)]
-pub enum Nsec3OptOut {
-    /// No opt-out. The opt-out flag of NSEC3 RRs will NOT be set and insecure
-    /// delegations will be included in the NSEC3 chain.
-    #[default]
-    NoOptOut,
-
-    /// Opt-out. The opt-out flag of NSEC3 RRs will be set and insecure
-    /// delegations will NOT be included in the NSEC3 chain.
-    OptOut,
-
-    /// Opt-out (flags only). The opt-out flag of NSEC3 RRs will be set and
-    /// insecure delegations will be included in the NSEC3 chain.
-    OptOutFlagsOnly,
-}
-
-// TODO: Add tests for nsec3s() that validate the following from RFC 5155:
-//
-// https://www.rfc-editor.org/rfc/rfc5155.html#section-7.1
-// 7.1. Zone Signing
-//     "Zones using NSEC3 must satisfy the following properties:
-//
-//      o  Each owner name within the zone that owns authoritative RRSets
-//         MUST have a corresponding NSEC3 RR.  Owner names that correspond
-//         to unsigned delegations MAY have a corresponding NSEC3 RR.
-//         However, if there is not a corresponding NSEC3 RR, there MUST be
-//         an Opt-Out NSEC3 RR that covers the "next closer" name to the
-//         delegation.  Other non-authoritative RRs are not represented by
-//         NSEC3 RRs.
-//
-//      o  Each empty non-terminal MUST have a corresponding NSEC3 RR, unless
-//         the empty non-terminal is only derived from an insecure delegation
-//         covered by an Opt-Out NSEC3 RR.
-//
-//      o  The TTL value for any NSEC3 RR SHOULD be the same as the minimum
-//         TTL value field in the zone SOA RR.
-//
-//      o  The Type Bit Maps field of every NSEC3 RR in a signed zone MUST
-//         indicate the presence of all types present at the original owner
-//         name, except for the types solely contributed by an NSEC3 RR
-//         itself.  Note that this means that the NSEC3 type itself will
-//         never be present in the Type Bit Maps."
-
-//------------ DesignatedSigningKey ------------------------------------------
-
-pub trait DesignatedSigningKey<Octs, Inner>:
-    Deref<Target = SigningKey<Octs, Inner>>
-where
-    Octs: AsRef<[u8]>,
-    Inner: SignRaw,
-{
-    /// Should this key be used to "sign one or more other authentication keys
-    /// for a given zone" (RFC 4033 section 2 "Key Signing Key (KSK)").
-    fn signs_keys(&self) -> bool;
-
-    /// Should this key be used to "sign a zone" (RFC 4033 section 2 "Zone
-    /// Signing Key (ZSK)").
-    fn signs_zone_data(&self) -> bool;
-}
-
-//------------ IntendedKeyPurpose --------------------------------------------
-
-/// The purpose of a DNSSEC key from the perspective of an operator.
-#[derive(Copy, Clone, Debug, Eq, PartialEq)]
-pub enum IntendedKeyPurpose {
-    /// A key that signs DNSKEY RRSETs.
-    ///
-    /// RFC9499 DNS Terminology:
-    /// 10. General DNSSEC
-    /// Key signing key (KSK): DNSSEC keys that "only sign the apex DNSKEY
-    ///   RRset in a zone." (Quoted from RFC6781, Section 3.1)
-    KSK,
-
-    /// A key that signs non-DNSKEY RRSETs.
-    ///
-    /// RFC9499 DNS Terminology:
-    /// 10. General DNSSEC
-    /// Zone signing key (ZSK): "DNSSEC keys that can be used to sign all the
-    /// RRsets in a zone that require signatures, other than the apex DNSKEY
-    /// RRset." (Quoted from RFC6781, Section 3.1) Also note that a ZSK is
-    /// sometimes used to sign the apex DNSKEY RRset.
-    ZSK,
-
-    /// A key that signs both DNSKEY and other RRSETs.
-    ///
-    /// RFC 9499 DNS Terminology:
-    /// 10. General DNSSEC
-    /// Combined signing key (CSK): In cases where the differentiation between
-    /// the KSK and ZSK is not made, i.e., where keys have the role of both
-    /// KSK and ZSK, we talk about a Single-Type Signing Scheme." (Quoted from
-    /// [RFC6781], Section 3.1) This is sometimes called a "combined signing
-    /// key" or "CSK". It is operational practice, not protocol, that
-    /// determines whether a particular key is a ZSK, a KSK, or a CSK.
-    CSK,
-
-    /// A key that is not currently used for signing.
-    ///
-    /// This key should be added to the zone but not used to sign any RRSETs.
-    Inactive,
-}
-
-//------------ DnssecSigningKey ----------------------------------------------
-
-/// A key that can be used for DNSSEC signing.
-///
-/// This type carries metadata that signals to a DNSSEC signer how this key
-/// should impact the zone to be signed.
-pub struct DnssecSigningKey<Octs, Inner: SignRaw> {
-    /// The key to use to make DNSSEC signatures.
-    key: SigningKey<Octs, Inner>,
-
-    /// The purpose for which the operator intends the key to be used.
-    ///
-    /// Defines explicitly the purpose of the key which should be used instead
-    /// of attempting to infer the purpose of the key (to sign keys and/or to
-    /// sign other records) by examining the setting of the Secure Entry Point
-    /// and Zone Key flags on the key (i.e. whether the key is a KSK or ZSK or
-    /// something else).
-    purpose: IntendedKeyPurpose,
-
-    _phantom: PhantomData<(Octs, Inner)>,
-}
-
-impl<Octs, Inner: SignRaw> DnssecSigningKey<Octs, Inner> {
-    /// Create a new [`DnssecSigningKey`] by assocating intent with a
-    /// reference to an existing key.
-    pub fn new(
-        key: SigningKey<Octs, Inner>,
-        purpose: IntendedKeyPurpose,
-    ) -> Self {
-        Self {
-            key,
-            purpose,
-            _phantom: Default::default(),
-        }
-    }
-
-    pub fn new_ksk(key: SigningKey<Octs, Inner>) -> Self {
-        Self {
-            key,
-            purpose: IntendedKeyPurpose::KSK,
-            _phantom: Default::default(),
-        }
-    }
-
-    pub fn new_zsk(key: SigningKey<Octs, Inner>) -> Self {
-        Self {
-            key,
-            purpose: IntendedKeyPurpose::ZSK,
-            _phantom: Default::default(),
-        }
-    }
-
-    pub fn new_csk(key: SigningKey<Octs, Inner>) -> Self {
-        Self {
-            key,
-            purpose: IntendedKeyPurpose::CSK,
-            _phantom: Default::default(),
-        }
-    }
-
-    pub fn new_inactive_key(key: SigningKey<Octs, Inner>) -> Self {
-        Self {
-            key,
-            purpose: IntendedKeyPurpose::Inactive,
-            _phantom: Default::default(),
-        }
-    }
-}
-
-impl<Octs, Inner> DnssecSigningKey<Octs, Inner>
-where
-    Octs: AsRef<[u8]>,
-    Inner: SignRaw,
-{
-    pub fn purpose(&self) -> IntendedKeyPurpose {
-        self.purpose
-    }
-
-    pub fn into_inner(self) -> SigningKey<Octs, Inner> {
-        self.key
-    }
-
-    // Note: This cannot be done as impl AsRef because AsRef requires that the
-    // lifetime of the returned reference be 'static, and we don't do impl Any
-    // as then the caller has to deal with Option or Result because the type
-    // might not impl DesignatedSigningKey.
-    pub fn as_designated_signing_key(
-        &self,
-    ) -> &dyn DesignatedSigningKey<Octs, Inner> {
-        self
-    }
-}
-
-//--- impl Deref
-
-impl<Octs, Inner> Deref for DnssecSigningKey<Octs, Inner>
-where
-    Octs: AsRef<[u8]>,
-    Inner: SignRaw,
-{
-    type Target = SigningKey<Octs, Inner>;
-
-    fn deref(&self) -> &Self::Target {
-        &self.key
-    }
-}
-
-//--- impl From
-
-impl<Octs, Inner> From<SigningKey<Octs, Inner>>
-    for DnssecSigningKey<Octs, Inner>
-where
-    Octs: AsRef<[u8]>,
-    Inner: SignRaw,
-{
-    fn from(key: SigningKey<Octs, Inner>) -> Self {
-        let public_key = key.public_key();
-        match (
-            public_key.is_secure_entry_point(),
-            public_key.is_zone_signing_key(),
-        ) {
-            (true, _) => Self::new_ksk(key),
-            (false, true) => Self::new_zsk(key),
-            (false, false) => Self::new_inactive_key(key),
-        }
-    }
-}
-
-//--- impl DesignatedSigningKey
-
-impl<Octs, Inner> DesignatedSigningKey<Octs, Inner>
-    for DnssecSigningKey<Octs, Inner>
-where
-    Octs: AsRef<[u8]>,
-    Inner: SignRaw,
-{
-    fn signs_keys(&self) -> bool {
-        matches!(
-            self.purpose,
-            IntendedKeyPurpose::KSK | IntendedKeyPurpose::CSK
-        )
-    }
-
-    fn signs_zone_data(&self) -> bool {
-        matches!(
-            self.purpose,
-            IntendedKeyPurpose::ZSK | IntendedKeyPurpose::CSK
-        )
-    }
-}
-
-//------------ Operations ----------------------------------------------------
-
-// TODO: Move nsecs() and nsecs3() out of SortedRecords and make them also
-// take an iterator. This allows callers to pass an iterator over Record
-// rather than force them to create the SortedRecords type (which for example
-// in the case of a Zone we wouldn't have, but may instead be able to get an
-// iterator over the Zone). Also move out the helper functions. Maybe put them
-// all into a Signer struct?
-
-pub trait SigningKeyUsageStrategy<Octs, Inner>
-where
-    Octs: AsRef<[u8]>,
-    Inner: SignRaw,
-{
-    const NAME: &'static str;
-
-    fn select_signing_keys_for_rtype(
-        candidate_keys: &[&dyn DesignatedSigningKey<Octs, Inner>],
-        rtype: Option<Rtype>,
-    ) -> HashSet<usize> {
-        if matches!(rtype, Some(Rtype::DNSKEY)) {
-            Self::filter_keys(candidate_keys, |k| k.signs_keys())
-        } else {
-            Self::filter_keys(candidate_keys, |k| k.signs_zone_data())
-        }
-    }
-
-    fn filter_keys(
-        candidate_keys: &[&dyn DesignatedSigningKey<Octs, Inner>],
-        filter: fn(&dyn DesignatedSigningKey<Octs, Inner>) -> bool,
-    ) -> HashSet<usize> {
-        candidate_keys
-            .iter()
-            .enumerate()
-            .filter_map(|(i, &k)| filter(k).then_some(i))
-            .collect::<HashSet<_>>()
-    }
-}
-
-pub struct DefaultSigningKeyUsageStrategy;
-
-impl<Octs, Inner> SigningKeyUsageStrategy<Octs, Inner>
-    for DefaultSigningKeyUsageStrategy
-where
-    Octs: AsRef<[u8]>,
-    Inner: SignRaw,
-{
-    const NAME: &'static str = "Default key usage strategy";
-}
-
-pub struct Signer<
-    Octs,
-    Inner,
-    KeyStrat = DefaultSigningKeyUsageStrategy,
-    Sort = DefaultSorter,
-> where
-    Octs: AsRef<[u8]>,
-    Inner: SignRaw,
-    KeyStrat: SigningKeyUsageStrategy<Octs, Inner>,
-    Sort: Sorter,
-{
-    _phantom: PhantomData<(Octs, Inner, KeyStrat, Sort)>,
-}
-
-impl<Octs, Inner, KeyStrat, Sort> Default
-    for Signer<Octs, Inner, KeyStrat, Sort>
-where
-    Octs: AsRef<[u8]>,
-    Inner: SignRaw,
-    KeyStrat: SigningKeyUsageStrategy<Octs, Inner>,
-    Sort: Sorter,
-{
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl<Octs, Inner, KeyStrat, Sort> Signer<Octs, Inner, KeyStrat, Sort>
-where
-    Octs: AsRef<[u8]>,
-    Inner: SignRaw,
-    KeyStrat: SigningKeyUsageStrategy<Octs, Inner>,
-    Sort: Sorter,
-{
-    pub fn new() -> Self {
-        Self {
-            _phantom: PhantomData,
-        }
-    }
-}
-
-impl<Octs, Inner, KeyStrat, Sort> Signer<Octs, Inner, KeyStrat, Sort>
-where
-    Octs: AsRef<[u8]> + From<Box<[u8]>> + OctetsFrom<Vec<u8>>,
-    Inner: SignRaw,
-    KeyStrat: SigningKeyUsageStrategy<Octs, Inner>,
-    Sort: Sorter,
-{
-    /// Sign a zone using the given keys.
-    ///
-    /// Returns the collection of RRSIG and (optionally) DNSKEY RRs that must be
-    /// added to the given records in order to DNSSEC sign them.
-    ///
-    /// The given records MUST be sorted according to [`CanonicalOrd`].
-    #[allow(clippy::type_complexity)]
-    pub fn sign<N>(
-        &self,
-        apex: &FamilyName<N>,
-        families: RecordsIter<'_, N, ZoneRecordData<Octs, N>>,
-        keys: &[&dyn DesignatedSigningKey<Octs, Inner>],
-        add_used_dnskeys: bool,
-    ) -> Result<Vec<Record<N, ZoneRecordData<Octs, N>>>, SigningError>
-    where
-        N: ToName + Display + Clone + PartialEq + CanonicalOrd + Send,
-        Octs: Clone + Send,
-    {
-        debug!("Signer settings: add_used_dnskeys={add_used_dnskeys}, strategy: {}", KeyStrat::NAME);
-
-        if keys.is_empty() {
-            return Err(SigningError::NoKeysProvided);
-        }
-
-        // Work with indices because SigningKey doesn't impl PartialEq so we
-        // cannot use a HashSet to make a unique set of them.
-
-        let dnskey_signing_key_idxs = KeyStrat::select_signing_keys_for_rtype(
-            keys,
-            Some(Rtype::DNSKEY),
-        );
-
-        let non_dnskey_signing_key_idxs =
-            KeyStrat::select_signing_keys_for_rtype(keys, None);
-
-        let keys_in_use_idxs: HashSet<_> = non_dnskey_signing_key_idxs
-            .iter()
-            .chain(dnskey_signing_key_idxs.iter())
-            .collect();
-
-        if keys_in_use_idxs.is_empty() {
-            return Err(SigningError::NoSuitableKeysFound);
-        }
-
-        // TODO: use log::log_enabled instead.
-        // See: https://github.com/NLnetLabs/domain/pull/465
-        if enabled!(Level::DEBUG) {
-            fn debug_key<Octs: AsRef<[u8]>, Inner: SignRaw>(
-                prefix: &str,
-                key: &SigningKey<Octs, Inner>,
-            ) {
-                debug!(
-                    "{prefix} with algorithm {}, owner={}, flags={} (SEP={}, ZSK={}) and key tag={}",
-                    key.algorithm()
-                        .to_mnemonic_str()
-                        .map(|alg| format!("{alg} ({})", key.algorithm()))
-                        .unwrap_or_else(|| key.algorithm().to_string()),
-                    key.owner(),
-                    key.flags(),
-                    key.is_secure_entry_point(),
-                    key.is_zone_signing_key(),
-                    key.public_key().key_tag(),
-                )
-            }
-
-            let num_keys = keys_in_use_idxs.len();
-            debug!(
-                "Signing with {} {}:",
-                num_keys,
-                if num_keys == 1 { "key" } else { "keys" }
-            );
-
-            for idx in &keys_in_use_idxs {
-                let key = keys[**idx];
-                let is_dnskey_signing_key =
-                    dnskey_signing_key_idxs.contains(idx);
-                let is_non_dnskey_signing_key =
-                    non_dnskey_signing_key_idxs.contains(idx);
-                let usage =
-                    if is_dnskey_signing_key && is_non_dnskey_signing_key {
-                        "CSK"
-                    } else if is_dnskey_signing_key {
-                        "KSK"
-                    } else if is_non_dnskey_signing_key {
-                        "ZSK"
-                    } else {
-                        "Unused"
-                    };
-                debug_key(&format!("Key[{idx}]: {usage}"), key);
-            }
-        }
-
-        let mut res: Vec<Record<N, ZoneRecordData<Octs, N>>> = Vec::new();
-        let mut buf = Vec::new();
-        let mut cut: Option<FamilyName<N>> = None;
-        let mut families = families.peekable();
-
-        // Are we signing the entire tree from the apex down or just some child records?
-        // Use the first found SOA RR as the apex. If no SOA RR can be found assume that
-        // we are only signing records below the apex.
-        let apex_ttl = families.peek().and_then(|first_family| {
-            first_family.records().find_map(|rr| {
-                if rr.owner() == apex.owner() && rr.rtype() == Rtype::SOA {
-                    if let ZoneRecordData::Soa(soa) = rr.data() {
-                        return Some(soa.minimum());
-                    }
-                }
-                None
-            })
-        });
-
-        if let Some(soa_minimum_ttl) = apex_ttl {
-            // Sign the apex
-            // SAFETY: We just checked above if the apex records existed.
-            let apex_family = families.next().unwrap();
-
-            let apex_rrsets = apex_family
-                .rrsets()
-                .filter(|rrset| rrset.rtype() != Rtype::RRSIG);
-
-            // Generate or extend the DNSKEY RRSET with the keys that we will sign
-            // apex DNSKEY RRs and zone RRs with.
-            let apex_dnskey_rrset = apex_family
-                .rrsets()
-                .find(|rrset| rrset.rtype() == Rtype::DNSKEY);
-
-            let mut augmented_apex_dnskey_rrs =
-                SortedRecords::<_, _, Sort>::new();
-
-            // Determine the TTL of any existing DNSKEY RRSET and use that as the
-            // TTL for DNSKEY RRs that we add. If none, then fall back to the SOA
-            // mininmum TTL.
-            //
-            // Applicable sections from RFC 1033:
-            //   TTL's (Time To Live)
-            //     "Also, all RRs with the same name, class, and type should
-            //      have the same TTL value."
-            //
-            //   RESOURCE RECORDS
-            //     "If you leave the TTL field blank it will default to the
-            //     minimum time specified in the SOA record (described
-            //     later)."
-            let dnskey_rrset_ttl = if let Some(rrset) = apex_dnskey_rrset {
-                let ttl = rrset.ttl();
-                augmented_apex_dnskey_rrs.extend(rrset.iter().cloned());
-                ttl
-            } else {
-                soa_minimum_ttl
-            };
-
-            for public_key in
-                keys_in_use_idxs.iter().map(|&&idx| keys[idx].public_key())
-            {
-                let dnskey = public_key.to_dnskey();
-
-                let signing_key_dnskey_rr = Record::new(
-                    apex.owner().clone(),
-                    apex.class(),
-                    dnskey_rrset_ttl,
-                    Dnskey::convert(dnskey.clone()).into(),
-                );
-
-                // Add the DNSKEY RR to the set of DNSKEY RRs to create RRSIGs for.
-                let is_new_dnskey = augmented_apex_dnskey_rrs
-                    .insert(signing_key_dnskey_rr)
-                    .is_ok();
-
-                if add_used_dnskeys && is_new_dnskey {
-                    // Add the DNSKEY RR to the set of new RRs to output for the zone.
-                    res.push(Record::new(
-                        apex.owner().clone(),
-                        apex.class(),
-                        dnskey_rrset_ttl,
-                        Dnskey::convert(dnskey).into(),
-                    ));
-                }
-            }
-
-            let augmented_apex_dnskey_rrset =
-                Rrset::new(augmented_apex_dnskey_rrs.as_slice());
-
-            // Sign the apex RRSETs in canonical order.
-            for rrset in apex_rrsets
-                .filter(|rrset| rrset.rtype() != Rtype::DNSKEY)
-                .chain(std::iter::once(augmented_apex_dnskey_rrset))
-            {
-                // For the DNSKEY RRSET, use signing keys chosen for that
-                // purpose and sign the augmented set of DNSKEY RRs that we
-                // have generated rather than the original set in the
-                // zonefile.
-                let signing_key_idxs = if rrset.rtype() == Rtype::DNSKEY {
-                    &dnskey_signing_key_idxs
-                } else {
-                    &non_dnskey_signing_key_idxs
-                };
-
-                for key in signing_key_idxs.iter().map(|&idx| keys[idx]) {
-                    // A copy of the family name. We’ll need it later.
-                    let name = apex_family.family_name().cloned();
-
-                    let rrsig_rr =
-                        Self::sign_rrset(key, &rrset, &name, apex, &mut buf)?;
-                    res.push(rrsig_rr);
-                    debug!(
-                        "Signed {} RRs in RRSET {} at the zone apex with keytag {}",
-                        rrset.iter().len(),
-                        rrset.rtype(),
-                        key.public_key().key_tag()
-                    );
-                }
-            }
-        }
-
-        // For all RRSETs below the apex
-        for family in families {
-            // If the owner is out of zone, we have moved out of our zone and
-            // are done.
-            if !family.is_in_zone(apex) {
-                break;
-            }
-
-            // If the family is below a zone cut, we must ignore it.
-            if let Some(ref cut) = cut {
-                if family.owner().ends_with(cut.owner()) {
-                    continue;
-                }
-            }
-
-            // A copy of the family name. We’ll need it later.
-            let name = family.family_name().cloned();
-
-            // If this family is the parent side of a zone cut, we keep the
-            // family name for later. This also means below that if
-            // `cut.is_some()` we are at the parent side of a zone.
-            cut = if family.is_zone_cut(apex) {
-                Some(name.clone())
-            } else {
-                None
-            };
-
-            for rrset in family.rrsets() {
-                if cut.is_some() {
-                    // If we are at a zone cut, we only sign DS and NSEC
-                    // records. NS records we must not sign and everything
-                    // else shouldn’t be here, really.
-                    if rrset.rtype() != Rtype::DS
-                        && rrset.rtype() != Rtype::NSEC
-                    {
-                        continue;
-                    }
-                } else {
-                    // Otherwise we only ignore RRSIGs.
-                    if rrset.rtype() == Rtype::RRSIG {
-                        continue;
-                    }
-                }
-
-                for key in
-                    non_dnskey_signing_key_idxs.iter().map(|&idx| keys[idx])
-                {
-                    let rrsig_rr =
-                        Self::sign_rrset(key, &rrset, &name, apex, &mut buf)?;
-                    res.push(rrsig_rr);
-                    debug!(
-                        "Signed {} RRSET at {} with keytag {}",
-                        rrset.rtype(),
-                        rrset.family_name().owner(),
-                        key.public_key().key_tag()
-                    );
-                }
-            }
-        }
-
-        debug!("Returning {} records from signing", res.len());
-
-        Ok(res)
-    }
-
-    fn sign_rrset<N, D>(
-        key: &SigningKey<Octs, Inner>,
-        rrset: &Rrset<'_, N, D>,
-        name: &FamilyName<N>,
-        apex: &FamilyName<N>,
-        buf: &mut Vec<u8>,
-    ) -> Result<Record<N, ZoneRecordData<Octs, N>>, SigningError>
-    where
-        N: ToName + Clone + Send,
-        D: RecordData
-            + ComposeRecordData
-            + From<Dnskey<Octs>>
-            + CanonicalOrd
-            + Send,
-    {
-        let (inception, expiration) = key
-            .signature_validity_period()
-            .ok_or(SigningError::KeyLacksSignatureValidityPeriod)?
-            .into_inner();
-        // RFC 4034
-        // 3.  The RRSIG Resource Record
-        //   "The TTL value of an RRSIG RR MUST match the TTL value of the
-        //    RRset it covers.  This is an exception to the [RFC2181] rules
-        //    for TTL values of individual RRs within a RRset: individual
-        //    RRSIG RRs with the same owner name will have different TTL
-        //    values if the RRsets they cover have different TTL values."
-        let rrsig = ProtoRrsig::new(
-            rrset.rtype(),
-            key.algorithm(),
-            name.owner().rrsig_label_count(),
-            rrset.ttl(),
-            expiration,
-            inception,
-            key.public_key().key_tag(),
-            // The fns provided by `ToName` state in their RustDoc that they
-            // "Converts the name into a single, uncompressed name" which
-            // matches the RFC 4034 section 3.1.7 requirement that "A sender
-            // MUST NOT use DNS name compression on the Signer's Name field
-            // when transmitting a RRSIG RR.".
-            //
-            // We don't need to make sure here that the signer name is in
-            // canonical form as required by RFC 4034 as the call to
-            // `compose_canonical()` below will take care of that.
-            apex.owner().clone(),
-        );
-        buf.clear();
-        rrsig.compose_canonical(buf).unwrap();
-        for record in rrset.iter() {
-            record.compose_canonical(buf).unwrap();
-        }
-        let signature = key.raw_secret_key().sign_raw(&*buf).unwrap();
-        let signature = signature.as_ref().to_vec();
-        let Ok(signature) = signature.try_octets_into() else {
-            return Err(SigningError::OutOfMemory);
-        };
-        let rrsig = rrsig.into_rrsig(signature).expect("long signature");
-        Ok(Record::new(
-            name.owner().clone(),
-            name.class(),
-            rrset.ttl(),
-            ZoneRecordData::Rrsig(rrsig),
-        ))
-    }
-}
-
-pub fn mk_hashed_nsec3_owner_name<N, Octs, SaltOcts>(
-    name: &N,
-    alg: Nsec3HashAlg,
-    iterations: u16,
-    salt: &Nsec3Salt<SaltOcts>,
-    apex_owner: &N,
-) -> Result<N, Nsec3HashError>
-where
-    N: ToName + From<Name<Octs>>,
-    Octs: FromBuilder,
-    <Octs as FromBuilder>::Builder: EmptyBuilder + AsRef<[u8]> + AsMut<[u8]>,
-    SaltOcts: AsRef<[u8]>,
-{
-    let base32hex_label =
-        mk_base32hex_label_for_name(name, alg, iterations, salt)?;
-    Ok(append_origin(base32hex_label, apex_owner))
-}
-
-fn append_origin<N, Octs>(base32hex_label: String, apex_owner: &N) -> N
-where
-    N: ToName + From<Name<Octs>>,
-    Octs: FromBuilder,
-    <Octs as FromBuilder>::Builder: EmptyBuilder + AsRef<[u8]> + AsMut<[u8]>,
-{
-    let mut builder = NameBuilder::<Octs::Builder>::new();
-    builder.append_label(base32hex_label.as_bytes()).unwrap();
-    let owner_name = builder.append_origin(apex_owner).unwrap();
-    let owner_name: N = owner_name.into();
-    owner_name
-}
-
-fn mk_base32hex_label_for_name<N, SaltOcts>(
-    name: &N,
-    alg: Nsec3HashAlg,
-    iterations: u16,
-    salt: &Nsec3Salt<SaltOcts>,
-) -> Result<String, Nsec3HashError>
-where
-    N: ToName,
-    SaltOcts: AsRef<[u8]>,
-{
-    let hash_octets: Vec<u8> =
-        nsec3_hash(name, alg, iterations, salt)?.into_octets();
-    Ok(base32::encode_string_hex(&hash_octets).to_ascii_lowercase())
-}
-
-//------------ Nsec3HashProvider ---------------------------------------------
-
-pub trait Nsec3HashProvider<N, Octs> {
-    fn get_or_create(
-        &mut self,
-        unhashed_owner_name: &N,
-    ) -> Result<N, Nsec3HashError>;
-}
-
-pub struct OnDemandNsec3HashProvider<N, SaltOcts> {
-    alg: Nsec3HashAlg,
-    iterations: u16,
-    salt: Nsec3Salt<SaltOcts>,
-    apex_owner: N,
-}
-
-impl<N, SaltOcts> OnDemandNsec3HashProvider<N, SaltOcts> {
-    pub fn new(
-        alg: Nsec3HashAlg,
-        iterations: u16,
-        salt: Nsec3Salt<SaltOcts>,
-        apex_owner: N,
-    ) -> Self {
-        Self {
-            alg,
-            iterations,
-            salt,
-            apex_owner,
-        }
-    }
-
-    pub fn algorithm(&self) -> Nsec3HashAlg {
-        self.alg
-    }
-
-    pub fn iterations(&self) -> u16 {
-        self.iterations
-    }
-
-    pub fn salt(&self) -> &Nsec3Salt<SaltOcts> {
-        &self.salt
-    }
-}
-
-impl<N, Octs, SaltOcts> Nsec3HashProvider<N, Octs>
-    for OnDemandNsec3HashProvider<N, SaltOcts>
-where
-    N: ToName + From<Name<Octs>>,
-    Octs: FromBuilder,
-    <Octs as FromBuilder>::Builder: EmptyBuilder + AsRef<[u8]> + AsMut<[u8]>,
-    SaltOcts: AsRef<[u8]>,
-{
-    fn get_or_create(
-        &mut self,
-        unhashed_owner_name: &N,
-    ) -> Result<N, Nsec3HashError> {
-        mk_hashed_nsec3_owner_name(
-            unhashed_owner_name,
-            self.alg,
-            self.iterations,
-            &self.salt,
-            &self.apex_owner,
-        )
     }
 }
